backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymarket-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # Web
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",

    # LangGraph / LangChain
    "langgraph>=0.2,<0.3",
    "langchain-core>=0.3,<0.4",

    # LLM providers
    "anthropic>=0.40",
    "httpx>=0.27",

    # Scheduler — pin 3.x, v4 has breaking API changes
    "apscheduler>=3.10,<4",

    # Utils
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "python-dotenv>=1.0",
    "tenacity>=8.3",

    # Vector DB + embeddings
    "qdrant-client>=1.9",
    "sentence-transformers>=3.0",
]

[project.optional-dependencies]
clob = [
    # Install separately when CLOB trading is needed:
    # pip install ".[clob]"
    "py-clob-client>=0.18",
]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
